Rename the module holding the Result more logically
