+++ src/lib.rs
@@ -36,8 +36,8 @@ pub mod data_types;
 pub(crate) use self::data_types::{unsafe_guid, Identify};
 pub use self::data_types::{CStr16, CStr8, Char16, Char8, Event, Guid, Handle};
 
-mod error;
-pub use self::error::{Completion, Result, ResultExt, Status};
+mod result;
+pub use self::result::{Completion, Result, ResultExt, Status};
 
 pub mod table;
