+++ src/result/completion.rs
